# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing dog care activities, reservations, passes and users. Protected endpoints require a Bearer JWT obtained from the login endpoint.",
        "title": "DogPaw API",
        "contact": {},
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/passes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of all passes in the system.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "List all passes (admin only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of passes to return (default 50, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of passes to skip for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of passes",
                        "schema": {
                            "$ref": "#/definitions/handler.listPassesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/passes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single pass by its id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "Get a pass by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pass ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pass found",
                        "schema": {
                            "$ref": "#/definitions/handler.passResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Pass not found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially updates a pass. Only price, pass_type, and\nexpires_at are editable. num_of_sessions,\nremaining_sessions, user_id, and created_at are\nimmutable to preserve the audit-log invariant. An\nempty body is a no-op.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "Patch a pass",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pass ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to patch",
                        "name": "pass",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.modifyPassRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated pass",
                        "schema": {
                            "$ref": "#/definitions/handler.passResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid id, body, or field value",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Pass not found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{user_id}/passes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of passes owned by the user identified by the user_id path param.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "List passes owned by a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Owner user ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of passes to return (default 50, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of passes to skip for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of passes",
                        "schema": {
                            "$ref": "#/definitions/handler.listPassesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user_id",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new prepaid session pack for the user identified by the user_id path param. Price is stored in cents.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "Register a new pass (bono) for a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Owner user ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pass to create",
                        "name": "pass",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.registerPassRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pass created",
                        "schema": {
                            "$ref": "#/definitions/handler.registerPassResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user_id, request body, or missing fields",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "Field 'breed' is required"
                },
                "error": {
                    "type": "string",
                    "example": "validation"
                },
                "field": {
                    "type": "string",
                    "example": "breed"
                }
            }
        },
...
        "handler.listPassesResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 3
                },
                "limit": {
                    "type": "integer",
                    "example": 50
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "passes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.passDTO"
                    }
                }
            }
        },
...
        "handler.modifyPassRequest": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2027-06-30T23:59:59Z"
                },
                "pass_type": {
                    "type": "string",
                    "example": "ESPECIFICO"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 15000
                }
            }
        },
        "handler.passDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-07-01T10:00:00Z"
                },
                "expires_at": {
                    "type": "string",
                    "example": "2026-12-31T23:59:59Z"
                },
                "id": {
                    "type": "integer",
                    "example": 42
                },
                "num_of_sessions": {
                    "type": "integer",
                    "example": 10
                },
                "pass_type": {
                    "type": "string",
                    "example": "GENERICO"
                },
                "price": {
                    "type": "integer",
                    "example": 12000
                },
                "remaining_sessions": {
                    "type": "integer",
                    "example": 7
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-07-15T14:30:00Z"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "handler.passResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-07-01T10:00:00Z"
                },
                "expires_at": {
                    "type": "string",
                    "example": "2026-12-31T23:59:59Z"
                },
                "id": {
                    "type": "integer",
                    "example": 42
                },
                "num_of_sessions": {
                    "type": "integer",
                    "example": 10
                },
                "pass_type": {
                    "type": "string",
                    "example": "GENERICO"
                },
                "price": {
                    "type": "integer",
                    "example": 12000
                },
                "remaining_sessions": {
                    "type": "integer",
                    "example": 7
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-07-15T14:30:00Z"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handler.registerPassRequest": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2026-12-31T23:59:59Z"
                },
                "num_of_sessions": {
                    "type": "integer",
                    "example": 10
                },
                "pass_type": {
                    "type": "string",
                    "example": "GENERICO"
                },
                "price": {
                    "type": "integer",
                    "example": 12000
                }
            }
        },
        "handler.registerPassResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Paste the token returned by /api/v1/auth/login as \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}